% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_rdd.R
\name{global_rdd}
\alias{global_rdd}
\title{Estimate A Global RDD}
\usage{
global_rdd(
  outcome,
  treatment,
  running_variable,
  c = 0,
  split_estimation = FALSE,
  return_graph = TRUE,
  cluster_id = NULL,
  covs = NULL,
  p_score_k = -1,
  estimator = "ATE",
  linear_penalty = FALSE,
  sp = NULL,
  linear_sp = -1,
  weights = NULL
)
}
\arguments{
\item{outcome}{A vector indicating the value of to the outcome of interest for each observation}

\item{treatment}{A vector indicating whether someone is treated (with a value of one) or not (with a value of zero) for each observation}

\item{running_variable}{A vector indicating the value of the running variable for each observation}

\item{c}{Where is the discontinuity in the running varaible?. Defaults to 0}

\item{return_graph}{Does the function return graphs plotting the results? Defaults to TRUE}

\item{cluster_id}{An optional argument indicating which cluster the different observations are in, for use in calculating the standard errors. Defaults to Null}

\item{covs}{Optional argument with list of column names that correspond to the covariates that are included. Defaults to NULL}

\item{p_score_k}{Optimal argument indicating the number of knots in the propensity score estimation. Default is -1, which means it's automatically selected.}

\item{estimator}{Do we return Conditional ATE or Conditional ATTs? Defaults to ATE}

\item{linear_penalty}{Do we shrink the linear terms as well? Defaults to FALSE, which means they are not shrunken.}

\item{sp}{What is the smoothing parameter? Defaults to Null, which means the smoothing parameters are automatically selected}

\item{linear_sp}{Do we shrink the linear terms as well? Defaults to FALSE, which means they are not shrunken.}

\item{weights}{An optional argument indicating the weights. Defaults to NULL which means each observation receives the same weight.}

\item{split_esitmation}{Do we randomly split the sample and estimate the propensity score and Global RDD on different samples?}
}
\value{
The code returns a variety of graphs illustrating the results (if return_graph == TRUE), measures of the ATE or ATT and its standard error as well as a table indicating the CATE or CATT and the standard errors. It also returns the output from teh estimated GAM and the the size of the first stage jump.
}
\description{
This function implements the Global RDD method, as described in Opper and Ozek (2024).
}
\keyword{GlobalRDD}
